rename component
