+++ frontend/src/components/App.jsx
@@ -3,12 +3,12 @@ import { connect } from 'react-redux';
 import { HashRouter, Route } from 'react-router-dom';
 
 // Components
-import EditorContainer from './EditorContainer';
+import EditorView from './EditorView';
 
 const App = () => {
   return (
     <HashRouter>
-      <Route path="/editor" component={EditorContainer} />
+      <Route path="/editor" component={EditorView} />
     </HashRouter>
   );
 };
