+++ frontend/src/components/EditorView.jsx
